cli/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dashborion-cli"
version = "0.1.0"
description = "Multi-cloud infrastructure dashboard CLI - Visualize and manage ECS, EKS, and CI/CD pipelines"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Kamorion", email = "contact@kamorion.com"}
]
maintainers = [
    {name = "Kamorion", email = "contact@kamorion.com"}
]
keywords = [
    "aws", "ecs", "eks", "kubernetes", "infrastructure",
    "dashboard", "cli", "devops", "monitoring",
    "codepipeline", "argocd", "diagram"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Monitoring",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
]
requires-python = ">=3.9"
dependencies = [
    "click>=8.0.0",
    "boto3>=1.26.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "kubernetes>=26.0.0",
    "diagrams>=0.23.0",
    "tabulate>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.0.270",
    "mypy>=1.0.0",
    "types-PyYAML",
    "types-requests",
]
confluence = [
    "atlassian-python-api>=3.36.0",
]
all = [
    "dashborion-cli[dev,confluence]",
]

[project.urls]
Homepage = "https://github.com/KamorionLabs/dashborion"
Documentation = "https://github.com/KamorionLabs/dashborion#readme"
Repository = "https://github.com/KamorionLabs/dashborion.git"
Issues = "https://github.com/KamorionLabs/dashborion/issues"
Changelog = "https://github.com/KamorionLabs/dashborion/blob/main/CHANGELOG.md"

[project.scripts]
dashborion = "dashborion.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["dashborion*"]

[tool.setuptools.package-data]
dashborion = ["py.typed"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.mypy_cache
    | \.pytest_cache
    | \.venv
    | build
    | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py39"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
addopts = "-v --tb=short"
